+++ R/read_hpa_spatial_data.R
@@ -74,6 +74,7 @@ get_data_file_list <- function() {
   #   stringr::str_remove("_?[A-Z]?\\.rda$") |>
   #   unique()
   c(
+    "acpr",
     "lhn",
     "phn",
     "mb21_poly",